Move store-only invariants from beacon_chain to store level
Move invariants 5 (execution payload), 6 (blob), and 8 (state cache)
from beacon_chain/src/invariants.rs to store/src/invariants.rs since
they only access store fields.

Also remove Display impl for InvariantViolation (Debug is sufficient)
and use Debug formatting in test assertions.
